+++ examples/selenium/rspec-capybara/app/controllers/vendors_controller.rb
@@ -1,0 +1,5 @@
+class VendorsController < ApplicationController
+  def index
+    @vendors = Vendor.all
+  end
+end