+++ examples/selenium/rspec-capybara/config/routes.rb
@@ -1,0 +1,3 @@
+Rails.application.routes.draw do
+  get '/vendors', to: 'vendors#index', as: 'vendors'
+end