+++ examples/selenium/rspec-capybara/spec/helpers/vendors_helper_spec.rb
@@ -1,0 +1,15 @@
+require 'rails_helper'
+
+# Specs in this file have access to a helper object that includes
+# the VendorsHelper. For example:
+#
+# describe VendorsHelper do
+#   describe "string concat" do
+#     it "concats two strings with spaces" do
+#       expect(helper.concat_strings("this","that")).to eq("this that")
+#     end
+#   end
+# end
+RSpec.describe VendorsHelper, type: :helper do
+
+end