+++ site/dat/docs/Changelog.md
@@ -11,6 +11,7 @@
  - fix percentiles values for cloud-based tests
  - support collection-based cloud tests
  - add stubs for `screenshoter` and `capturehar` services
+ - add RSpec tests runner for Selenium
 
 ## 1.6.8 <sup>1 sep 2016</sup>
  - fix hamcrest installation for Java-based Selenium tests